refactor: updated aws connection module file naming
